Worked on cleaning up types that are different sizes on 32-bit and 64-bit platforms: Specifically, I got rid of a lot of 'long' and 'long long' in favor of int32 and int64 (etc.) versions defined in class.h
Don't forget to remove the '.so' files from the autotest directories
before re-running the tests! 



git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@2324 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ climate/csv_reader.cpp
@@ -197,7 +197,7 @@ int csv_reader::read_prop(char *line){ // already pulled the '$' off the front
 //		gl_error("csv_reader::read_prop ~ unable to set property \'%s\' to \'%s\'", propstr, valstr);
 //		return 0;
 //	}	
-	void *addr = (void *)((unsigned long long int)this + (unsigned long long int)prop->addr);
+	void *addr = (void *)((uint64)this + (uint64)prop->addr);
 	if(prop->ptype == PT_double){
 		if(1 != sscanf(valstr, "%lg", addr)){
 			gl_error("csv_reader::read_prop ~ unable to set property \'%s\' to \'%s\'", propstr, valstr);
@@ -339,7 +339,7 @@ int csv_reader::read_line(char *line){
 
 	while((token=strtok(NULL, ",\n\r")) != 0 && col < column_ct){
 		if(columns[col]->ptype == PT_double){
-			double *dptr = (double *)((unsigned long long int)(columns[col]->addr) + (unsigned long long int)(sample));
+			double *dptr = (double *)((uint64)(columns[col]->addr) + (uint64)(sample));
 			if(sscanf(token, "%lg", dptr) != 1){
 				gl_error("csv_reader::read_line ~ unable to set value \'%s\' to double property \'%s\'", token, columns[col]->name);
 				/* TROUBLESHOOT
